Handle language delete

+++ src/modals/LanguageDelete.vue
@@ -7,16 +7,21 @@
 				<div class="modal-title h5">{{ $t('modal.deleteLanguage') }}</div>
 			</div>
 			<div class="modal-body">
-				<div class="content">
+				<div v-if="!languageInUse" class="content">
 					<p>{{ $t('text.reallyDeleteLanguage', { name: languageName }) }}</p>
 					<p>{{ $t('text.cannotBeUndone') }}</p>
+				</div>
+				<div v-else class="content">
+					<p>{{ $t('text.languageInUse') }}</p>
 				</div>
 			</div>
 			<div class="modal-footer">
 				<a class="btn btn-link btn-gray" href="#" aria-label="Cancel" @click.prevent="$emit('closed')">
 					{{ $t('button.cancel') }}
 				</a>
-				<button class="btn btn-error ml-2" @click="deleteLanguage">{{ $t('button.delete') }}</button>
+				<button class="btn btn-error ml-2" :class="{ disabled: languageInUse }" @click="deleteLanguage">
+					{{ $t('button.delete') }}
+				</button>
 			</div>
 		</div>
 	</div>
@@ -28,19 +33,39 @@ export default {
 	props: {
 		active: Boolean,
 		languageName: String,
-		languageKey: String
+		languageKey: String,
+		songs: Object
 	},
 	methods: {
 		deleteLanguage () {
-			this.$db.collection('languages').doc(this.languageKey).delete().then(() => {
-				this.$emit('closed');
-				// toast success message
-				this.$notify({
-					title: this.$parent.$t('toast.languageDeleted'),
-					text: this.$parent.$t('toast.languageDeletedText'),
-					type: 'primary'
+			if (!this.languageInUse) {
+				this.$db.collection('languages').doc(this.languageKey).delete().then(() => {
+					this.$emit('closed');
+					// toast success message
+					this.$notify({
+						title: this.$parent.$t('toast.languageDeleted'),
+						text: this.$parent.$t('toast.languageDeletedText'),
+						type: 'primary'
+					});
+				}).catch((error) => this.throwError(error));
+			} else {
+				this.throwError({
+					code: this.$parent.$t('toast.languageInUse'),
+					message: this.$parent.$t('text.languageInUse')
 				});
-			}).catch((error) => this.throwError(error));
+			}
+		}
+	},
+	computed: {
+		languageInUse () {
+			for (const id in this.songs) {
+				if (Object.hasOwnProperty.call(this.songs, id)) {
+					if (this.songs[id].language == this.languageKey) {
+						return true;
+					};
+				}
+			}
+			return false;
 		}
 	}
 }
